Revert "Fixes #247"
This reverts commit d2fbe7b459f777220093110e4c574fb31770cb62.

+++ spring-security-rest/src/main/groovy/grails/plugin/springsecurity/rest/RestAuthenticationProvider.groovy
@@ -54,7 +54,7 @@ class RestAuthenticationProvider implements AuthenticationProvider {
             log.debug "Trying to validate token ${authenticationRequest.accessToken}"
             UserDetails userDetails = tokenStorageService.loadUserByToken(authenticationRequest.accessToken) as UserDetails
 
-            Long expiration = null
+            Integer expiration = null
             if (useJwt) {
                 Date now = new Date()
                 JWT jwt = jwtService.parse(authenticationRequest.accessToken)
